+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Galactic War setup is much more responsive as the galaxy is now only generated once you click "Go To War" and not every time you change a setting
+- Buff Catalyst Upgrade Tech
 
 ### Fixed