+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Structure Armor Tech not including the Halley in its tooltip
 - Orbital Engine Tech tooltip including structures
 - Bot Combat Tech health and speed bonuses not being applied to the Gil-E
+- Complete Bot Tech tooltip not listing the Stinger or Gil-E
+- Complete Orbital Tech tooltip not listing the ARKYD
 
 ## v5.26.1 - 2021-12-08